Fix return type for lpush and rpush so they provide a numeric return (length of list) for the value specified by redis.
Note: The current code may have a type inconsistency in that the native return from redis is an unsigned integer. However, because .Machine$integer.max is signed I don't coerce the response into an integer and instead allow it to return as a numeric. I'm not 100% of the sure of the consequences of this decision.